Support remote snapshotter to speed up image pulling
Among container's lifecycle, pulling image is one of the biggest performance
bottleneck on container startup processes.
One research shows that time for pulling accounts for 76% of container startup
time[FAST '16](https://www.usenix.org/node/194431).

Remote snapshotter is one of the solutions of the issue. And discussed on some
issue threads(#2943, etc.). This implementation partially based on the
discussion but slightly different to make it work with metadata snapshotter
which binds each snapshots to namespaces.

Signed-off-by: Kohei Tokunaga <kouhei.tokunaga.uy@hco.ntt.co.jp>